Move stg code up a level
