--- src/eval/common/saturate.rs
@@ -1,1 +1,0 @@
-//! Ensure all calls to intrinsics are saturate (no partials)
