+++ src/eval/mod.rs
@@ -1,6 +1,6 @@
 pub mod emit;
 pub mod error;
 pub mod intrinsics;
-pub mod machines;
 pub mod primitive;
+pub mod stg;
 pub mod types;
